dispatch-bot-ai/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Environment and configuration
python-dotenv==1.0.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Production server
gunicorn==21.2.0

# External API integrations  
openai==1.3.7          # OpenAI GPT-4 API client
googlemaps==4.10.0     # Google Maps API client
requests==2.31.0       # HTTP client for API calls

# Optional dependencies for future features
# redis==5.0.1      # For caching if needed
# sqlalchemy==2.0.23 # For database if needed